[MIG] product_reference_required to 11.0 [FIX] readme [MOV] product to product_template and  product_product [REM] dead lines and data tag
