+++ product_reference_required/data/default_code_data.xml
@@ -1,16 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data noupdate="1">
-
-        <!--<record id="sequence_type" model="ir.sequence.type">-->
-            <!--<field name="name">Product Default Code</field>-->
-            <!--<field name="code">product.default_code</field>-->
-        <!--</record>-->
-
-        <record id="sequence" model="ir.sequence">
-            <field name="name">Product Default Code</field>
-            <field name="code">product.default_code</field>
-        </record>
-
-    </data>
+<odoo noupdate="1">
+    <record id="sequence" model="ir.sequence">
+        <field name="name">Product Default Code</field>
+        <field name="code">product.default_code</field>
+    </record>
 </odoo>
