+++ product_reference_required/demo/default_code_sequence_demo.xml
@@ -1,12 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
-<openerp>
-    <data noupdate="1">
-
-        <record id="sequence" model="ir.sequence">
-            <field name="name">Default Code Demo</field>
-            <field name="code">product.default_code</field>
-            <field name="padding">4</field>
-        </record>
-
-    </data>
+<odoo noupdate="1">
+    <record id="sequence" model="ir.sequence">
+        <field name="name">Default Code Demo</field>
+        <field name="code">product.default_code</field>
+        <field name="padding">4</field>
+    </record>
 </odoo>
