+++ product_reference_required/models/__init__.py
@@ -3,4 +3,5 @@
 # directory
 ##############################################################################
 
-from . import product
+from . import product_product
+from . import product_template
